refactor (dashboard): change test config to avoid errors
